Don't auto-populate products or location types
